requirements.txt
pyyaml
python-resize-image
